[ADD] user service

+++ class-utility/src/main/java/org/example/response/HttpResponse.java
@@ -60,7 +60,7 @@ public class HttpResponse {
     @Setter
     @Getter
     public static class PaginationParam{
-        private Integer offset;
+        private Long offset;
         private Integer limit;
         private QueryParam param;
     }
